+++ packages/x_storage_flutter/lib/x_storage_flutter.dart
@@ -1,4 +1,2 @@
-library x_storage_flutter;
-
 export 'src/x_storage_audio_service.dart';
 export 'src/x_storage_image.dart';