+++ packages/x_storage_flutter/test/x_storage_flutter_test.dart
@@ -1,5 +1,8 @@
 import 'package:flutter_test/flutter_test.dart';
-
 import 'package:x_storage_flutter/x_storage_flutter.dart';
 
-void main() {}
+void main() {
+  test('exports XStorageImage', () {
+    expect(XStorageImage, isNotNull);
+  });
+}